Move NetworkHandlerService to new package "network"
