+++ server/src/main/java/com/amnesica/belugaproject/services/aircraft/AirplanesLiveService.java
@@ -6,8 +6,8 @@ import com.amnesica.belugaproject.config.FeederMapping;
 import com.amnesica.belugaproject.entities.aircraft.RemoteAircraft;
 import com.amnesica.belugaproject.repositories.aircraft.RemoteAircraftRepository;
 import com.amnesica.belugaproject.services.data.AircraftDataService;
-import com.amnesica.belugaproject.services.helper.NetworkHandlerService;
 import com.amnesica.belugaproject.services.helper.Request;
+import com.amnesica.belugaproject.services.network.NetworkHandlerService;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONArray;
 import org.json.JSONObject;
