+++ server/src/main/java/com/amnesica/belugaproject/services/aircraft/LocalFeederService.java
@@ -8,7 +8,7 @@ import com.amnesica.belugaproject.entities.data.AirportData;
 import com.amnesica.belugaproject.repositories.aircraft.AircraftRepository;
 import com.amnesica.belugaproject.services.data.AircraftDataService;
 import com.amnesica.belugaproject.services.data.AirportDataService;
-import com.amnesica.belugaproject.services.helper.NetworkHandlerService;
+import com.amnesica.belugaproject.services.network.NetworkHandlerService;
 import com.amnesica.belugaproject.services.trails.AircraftTrailService;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONArray;
