+++ server/src/main/java/com/amnesica/belugaproject/services/aircraft/OpenskyService.java
@@ -8,9 +8,9 @@ import com.amnesica.belugaproject.entities.trails.AircraftTrail;
 import com.amnesica.belugaproject.repositories.aircraft.RemoteAircraftRepository;
 import com.amnesica.belugaproject.services.data.AircraftDataService;
 import com.amnesica.belugaproject.services.helper.HelperService;
-import com.amnesica.belugaproject.services.helper.NetworkHandlerService;
 import com.amnesica.belugaproject.services.helper.Request;
 import com.amnesica.belugaproject.services.helper.TrailHelperService;
+import com.amnesica.belugaproject.services.network.NetworkHandlerService;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONArray;
 import org.json.JSONObject;
