+++ server/src/main/java/com/amnesica/belugaproject/services/aircraft/SpacecraftService.java
@@ -6,8 +6,8 @@ import com.amnesica.belugaproject.entities.aircraft.AircraftSuperclass;
 import com.amnesica.belugaproject.entities.aircraft.Spacecraft;
 import com.amnesica.belugaproject.repositories.aircraft.SpacecraftRepository;
 import com.amnesica.belugaproject.services.helper.HelperService;
-import com.amnesica.belugaproject.services.helper.NetworkHandlerService;
 import com.amnesica.belugaproject.services.helper.Request;
+import com.amnesica.belugaproject.services.network.NetworkHandlerService;
 import com.amnesica.belugaproject.services.trails.SpacecraftTrailService;
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
